Added basic config for ghostty
I decided to bite the bullet and try new tooling in 2025! It was a
fortunate chain of events. I was exploring ways of better adding AI into
my workflow, which led me to a Neovim-only plugin, which led me to
playing around with Neovim install, which led me to hearing about
Ghostty... and this is the only part that I'm fully satisfied with the
configuration so far!

But so, to keep things going, here's the first version of my Ghostty
config. I'm sure I'll keep tweaking!

Includes:
 - Colorscheme (Nord, which is a good base for Nordic in Neovim)
 - Some keybindings to more easily use arrows in the terminal (never
   leave the home row)
 - Use Fira Code (I'm keeping that one from 2024, but I'm adding the
   Nerd version of it, to have nice little icons)