Fixed rspec automation
For the moment the best is to stick with the old way of doing things
in a tmux pane. I've seen that there are a number of other test
runners, in the background or with notifications, like neotest, but
since I don't want to spend time troubleshooting, this is the simplest
and fastest way to get it running.